update schema path
