+++ app/schemas/base.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+from datetime import datetime
+
+
+class AppBaseModel(BaseModel):
+    created_at: datetime = None
+    updated_at: datetime = None
